requirements.txt
# 浏览器自动化
playwright==1.40.0

# OpenAI
openai==1.3.0

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 定时任务
schedule==1.2.0

# Windows通知
win10toast==0.9

# 图片处理
Pillow==10.1.0

# 异步HTTP
aiohttp==3.9.0

# 工具库
python-dateutil==2.8.2
colorlog==6.8.0

# 数据处理
ujson==5.8.0
